wip: save bf16 bitdecoding baseline before 2-bit work

+++ csrc/bit_decode/decode_api.cpp
@@ -194,23 +194,23 @@ void run_mha_fwd(Flash_fwd_params &params, cudaStream_t stream, bool force_split
     //     });
     // });
     if (params.num_splits <= 1 && !force_split_kernel) {  // If we don't set it num_splits == 0
-        run_mha_fwd_<cutlass::half_t, 128, false>(params, stream);
+        run_mha_fwd_<cutlass::bfloat16_t, 128, false>(params, stream);
     } else {
         if (params.quant_mode == "k-channel") {
             if (params.group_size == 128) {
-                run_mha_fwd_splitkv_dispatch<cutlass::half_t, 128, false, 1, num_bits, 128>(params, stream);
+                run_mha_fwd_splitkv_dispatch<cutlass::bfloat16_t, 128, false, 1, num_bits, 128>(params, stream);
             } else if (params.group_size == 64) {
-                // run_mha_fwd_splitkv_dispatch<cutlass::half_t, 128, false, 1, num_bits, 64>(params, stream);
+                // run_mha_fwd_splitkv_dispatch<cutlass::bfloat16_t, 128, false, 1, num_bits, 64>(params, stream);
             } else if (params.group_size == 32) {
-                run_mha_fwd_splitkv_dispatch<cutlass::half_t, 128, false, 1, num_bits, 32>(params, stream);
+                run_mha_fwd_splitkv_dispatch<cutlass::bfloat16_t, 128, false, 1, num_bits, 32>(params, stream);
             }
         } else {
             if (params.group_size == 128) {
-                // run_mha_fwd_splitkv_dispatch<cutlass::half_t, 128, false, 0, num_bits, 128>(params, stream);
+                // run_mha_fwd_splitkv_dispatch<cutlass::bfloat16_t, 128, false, 0, num_bits, 128>(params, stream);
             } else if (params.group_size == 64) {
-                // run_mha_fwd_splitkv_dispatch<cutlass::half_t, 128, false, 0, num_bits, 64>(params, stream);
+                // run_mha_fwd_splitkv_dispatch<cutlass::bfloat16_t, 128, false, 0, num_bits, 64>(params, stream);
             } else if (params.group_size == 32) {
-                // run_mha_fwd_splitkv_dispatch<cutlass::half_t, 128, false, 0, num_bits, 32>(params, stream);
+                // run_mha_fwd_splitkv_dispatch<cutlass::bfloat16_t, 128, false, 0, num_bits, 32>(params, stream);
             }
         }
     }
@@ -220,19 +220,19 @@ template <int num_bits>
 void run_kvcache_qpack(Flash_fwd_params &params, cudaStream_t stream) {
     if (params.quant_mode == "k-channel") {
         if (params.group_size == 32) {
-            run_kvcache_qpack_<cutlass::half_t, 128, 1, num_bits, 32>(params, stream);
+            run_kvcache_qpack_<cutlass::bfloat16_t, 128, 1, num_bits, 32>(params, stream);
         } else if (params.group_size == 64) {
-            // run_kvcache_qpack_<cutlass::half_t, 128, 1, num_bits, 64>(params, stream);
+            // run_kvcache_qpack_<cutlass::bfloat16_t, 128, 1, num_bits, 64>(params, stream);
         } else if (params.group_size == 128) {
-            run_kvcache_qpack_<cutlass::half_t, 128, 1, num_bits, 128>(params, stream);
+            run_kvcache_qpack_<cutlass::bfloat16_t, 128, 1, num_bits, 128>(params, stream);
         }
     } else {
         if (params.group_size == 32) {
-            // run_kvcache_qpack_<cutlass::half_t, 128, 0, num_bits, 32>(params, stream);
+            // run_kvcache_qpack_<cutlass::bfloat16_t, 128, 0, num_bits, 32>(params, stream);
         } else if (params.group_size == 64) {
-            // run_kvcache_qpack_<cutlass::half_t, 128, 0, num_bits, 64>(params, stream);
+            // run_kvcache_qpack_<cutlass::bfloat16_t, 128, 0, num_bits, 64>(params, stream);
         } else if (params.group_size == 128) {
-            // run_kvcache_qpack_<cutlass::half_t, 128, 0, num_bits, 128>(params, stream);
+            // run_kvcache_qpack_<cutlass::bfloat16_t, 128, 0, num_bits, 128>(params, stream);
         }
     }
 }
